first pass at monad-inspired hash parsing
This is a fresh look at an actual problem: how Updraft parses its
incoming document creation requests.  Presently this is tackled in a
rather ad-hoc way, in that there are a number of underspecified hashes
and we sometimes get validation errors in unexpected places.

The approach here is:

* "Parse, don't validate" - everywhere that we check an input is
well-specified, we also convert the external representation into the
object we actually need

* _inspired_ by monads, but is possibly not actually compliant with
the monad laws. (Yes, this is why we have the rather opaque method
names `unit` and `lift`)

The biggest ugliness in this right now is that the Result type only
really accommodated key->value maps, and sometimes we'd like a value
that's just an array. You'll see in Template.parse that we have to
fake some field names. This is temporary and will be addressed once
I think of a nice way to do so

There is also some work that could be done to make the syntax sweeter.

Start by reading spec/stitchup/template_spec.rb to see how it hangs
together